requirements.txt
Django>=6.0,<6.1
django-allauth
django-tailwind
django-htmx
django-crispy-forms
crispy-tailwind
python-dotenv
whitenoise
gunicorn
stripe
dj-database-url
psycopg[binary]
django-tenants
Pillow
ruff